% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_carhyce_stations.R
\docType{data}
\name{carhyce_stations}
\alias{carhyce_stations}
\title{carhyce_stations}
\format{
A data frame with 2707 rows and 3 variables:
\describe{
\item{ code_station }{  character }
\item{ name_station }{  character }
\item{ geometry }{  sfc_POINT,sfc }
}
}
\source{
Source
}
\usage{
carhyce_stations
}
\description{
Description.
}
\keyword{datasets}
