% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Format column names}
\usage{
clean_column_names(names)
}
\arguments{
\item{names}{the column names to format.}
}
\value{
A data frame with modified column names.
}
\description{
Format column names
}
