% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{clean_duplicated}
\alias{clean_duplicated}
\title{Drop all the duplicated rows from a duplicated data.frame.}
\usage{
clean_duplicated(
  dataset,
  duplicated_dataset,
  axis_field = "AXIS",
  measure_field = "M"
)
}
\arguments{
\item{dataset}{data.frame.}

\item{duplicated_dataset}{data.frame produce by check_duplicate function.}

\item{axis_field}{axis stream id.}

\item{measure_field}{axis stream measure.}
}
\value{
sf data.frame.
}
\description{
Drop all the duplicated rows from a duplicated data.frame.
}
