% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_continuity.R
\docType{data}
\name{continuity}
\alias{continuity}
\title{continuity}
\format{
A data frame with 9680 rows and 9 variables:
\describe{
\item{ swath }{  numeric }
\item{ label }{  character }
\item{ side }{  character }
\item{ area }{  numeric }
\item{ width1 }{  numeric }
\item{ width2 }{  numeric }
\item{ axis }{  numeric }
\item{ measure }{  numeric }
\item{ width_continuity }{  numeric }
}
}
\source{
Source
}
\usage{
continuity
}
\description{
Description.
}
\keyword{datasets}
