% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_continuity_data.R
\docType{data}
\name{continuity_data}
\alias{continuity_data}
\title{continuity_data}
\format{
A data frame with 5412 rows and 8 variables:
\describe{
\item{ X }{  integer }
\item{ measure }{  integer }
\item{ continuity }{  character }
\item{ continuity_area }{  integer }
\item{ continuity_width }{  numeric }
\item{ continuity_weighted_area }{  integer }
\item{ continuity_weighted_width }{  numeric }
\item{ AXIS }{  integer }
}
}
\source{
Source
}
\usage{
continuity_data
}
\description{
Description.
}
\keyword{datasets}
