% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_view.R
\name{continuity_full_side_view_geom}
\alias{continuity_full_side_view_geom}
\title{Create continuity_full_side_view_geom view}
\usage{
continuity_full_side_view_geom(
  db_con,
  view_name = "continuity_area_full_side_geom"
)
}
\arguments{
\item{db_con}{database connection parameters.}

\item{view_name}{view name.}
}
\value{
text
}
\description{
Create continuity_full_side_view_geom view
}
