% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_view.R
\name{continuity_width_full_side_view_geom}
\alias{continuity_width_full_side_view_geom}
\title{Create continuity_width_full_side_geom view}
\usage{
continuity_width_full_side_view_geom(
  db_con,
  view_name = "continuity_width_full_side_geom",
  continuity_table_name
)
}
\arguments{
\item{db_con}{database connection parameters.}

\item{view_name}{view name.}
}
\value{
text
}
\description{
Create continuity_width_full_side_geom view
}
