% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_width.R
\name{create_continuity_width_full_side_matview}
\alias{create_continuity_width_full_side_matview}
\title{Create continuity_width_full_side view}
\usage{
create_continuity_width_full_side_matview(
  db_con,
  view_name = "continuity_width_full_side"
)
}
\arguments{
\item{db_con}{database connection parameters.}

\item{view_name}{view name.}
}
\value{
text
}
\description{
Create continuity_width_full_side view
}
