% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_landcover_area.R
\name{create_landcover_area_full_side_view}
\alias{create_landcover_area_full_side_view}
\title{Create landcover_area_full_side view}
\usage{
create_landcover_area_full_side_view(
  db_con,
  view_name = "landcover_area_full_side"
)
}
\arguments{
\item{db_con}{database connection parameters.}

\item{view_name}{view name.}
}
\value{
text
}
\description{
Create landcover_area_full_side view
}
