% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_axis.R
\name{create_network_axis_view}
\alias{create_network_axis_view}
\title{Create network axis view for mapdoapp application.}
\usage{
create_network_axis_view(db_con, view_name = "network_axis")
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{view_name}{view name.}
}
\value{
text
}
\description{
Create network axis view for mapdoapp application.
}
