% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_swaths.R
\name{create_network_metrics_view}
\alias{create_network_metrics_view}
\title{Create network metrics view for mapdoapp application.}
\usage{
create_network_metrics_view(db_con, view_name)
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{view_name}{view name.}
}
\value{
text
}
\description{
Create network metrics view for mapdoapp application.
}
