% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bassin_hydrographique.R
\name{create_table_bassin_hydrographique}
\alias{create_table_bassin_hydrographique}
\title{Create bassin_hydrographique table structure.}
\usage{
create_table_bassin_hydrographique(
  table_name = "bassin_hydrographique",
  db_con
)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create bassin_hydrographique table structure.
}
