% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_area.R
\name{create_table_continuity_area}
\alias{create_table_continuity_area}
\title{Create continuity_area table structure.}
\usage{
create_table_continuity_area(table_name = "continuity_area", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create continuity_area table structure.
}
