% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_width.R
\name{create_table_continuity_width}
\alias{create_table_continuity_width}
\title{Create continuity_width table structure.}
\usage{
create_table_continuity_width(table_name = "continuity_width", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create continuity_width table structure.
}
