% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_elevation_profiles.R
\name{create_table_elevation_profiles}
\alias{create_table_elevation_profiles}
\title{Create elevation_profiles table structure.}
\usage{
create_table_elevation_profiles(table_name = "elevation_profiles", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create elevation_profiles table structure.
}
