% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_axis.R
\name{create_table_hydro_axis}
\alias{create_table_hydro_axis}
\title{Create hydro_axis table structure.}
\usage{
create_table_hydro_axis(table_name = "hydro_axis", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create hydro_axis table structure.
}
