% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_sites.R
\name{create_table_hydro_sites}
\alias{create_table_hydro_sites}
\title{Create hydro_sites table structure.}
\usage{
create_table_hydro_sites(table_name = "hydro_sites", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create hydro_sites table structure.
}
