% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_stations.R
\name{create_table_hydro_stations}
\alias{create_table_hydro_stations}
\title{Create hydro_stations table structure.}
\usage{
create_table_hydro_stations(table_name = "hydro_stations", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create hydro_stations table structure.
}
