% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_structure.R
\name{create_table_hydro_swaths}
\alias{create_table_hydro_swaths}
\title{Create hydro_swaths table structure}
\usage{
create_table_hydro_swaths(table_name = "hydro_swaths")
}
\arguments{
\item{table_name}{table name.}
}
\value{
text
}
\description{
Create hydro_swaths table structure
}
