% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_landcover.R
\name{create_table_landcover_area}
\alias{create_table_landcover_area}
\title{Create landcover_area table structure.}
\usage{
create_table_landcover_area(table_name = "landcover_area", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create landcover_area table structure.
}
