% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_structure.R
\name{create_table_public.region_hydrographique}
\alias{create_table_public.region_hydrographique}
\title{Create region_hydrographique table structure}
\usage{
create_table_public.region_hydrographique(
  table_name = "public.region_hydrographique"
)
}
\arguments{
\item{table_name}{table name.}
}
\value{
text
}
\description{
Create region_hydrographique table structure
}
