% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_structure.R
\name{create_table_region_hydrographique}
\alias{create_table_region_hydrographique}
\title{Create region_hydrographique table structure}
\usage{
create_table_region_hydrographique(
  table_name = "region_hydrographique",
  db_con
)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create region_hydrographique table structure
}
