% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_structure.R
\name{create_table_roe}
\alias{create_table_roe}
\title{Create roe table structure}
\usage{
create_table_roe(table_name = "roe", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create roe table structure
}
