% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_talweg_metrics.R
\name{create_table_talweg_metrics}
\alias{create_table_talweg_metrics}
\title{Create talweg_metrics table structure.}
\usage{
create_table_talweg_metrics(table_name = "talweg_metrics", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create talweg_metrics table structure.
}
