% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_valley_bottom.R
\name{create_table_valley_bottom}
\alias{create_table_valley_bottom}
\title{Create valley_bottom table structure.}
\usage{
create_table_valley_bottom(table_name = "valley_bottom", db_con)
}
\arguments{
\item{table_name}{table name.}

\item{db_con}{DBI database connection.}
}
\value{
text
}
\description{
Create valley_bottom table structure.
}
