% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_valley_bottom.R
\name{create_valley_bottom_full_side_matview}
\alias{create_valley_bottom_full_side_matview}
\title{Create valley_bottom_full_side view}
\usage{
create_valley_bottom_full_side_matview(
  db_con,
  view_name = "valley_bottom_full_side"
)
}
\arguments{
\item{db_con}{database connection parameters.}

\item{view_name}{view name.}
}
\value{
text
}
\description{
Create valley_bottom_full_side view
}
