% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_elevation_profiles.R
\docType{data}
\name{elevation_profiles}
\alias{elevation_profiles}
\title{elevation_profiles}
\format{
A data frame with 8800 rows and 9 variables:
\describe{
\item{ sample }{  numeric }
\item{ side }{  character }
\item{ quantile }{  numeric }
\item{ density }{  numeric }
\item{ mean }{  numeric }
\item{ profile }{  numeric }
\item{ axis }{  numeric }
\item{ measure }{  numeric }
\item{ distance }{  numeric }
}
}
\source{
Source
}
\usage{
elevation_profiles
}
\description{
Description.
}
\keyword{datasets}
