% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_width.R
\name{fct_continuity_width_insert_delete_reaction}
\alias{fct_continuity_width_insert_delete_reaction}
\title{Add trigger function to react from continuity_width insert or delete.}
\usage{
fct_continuity_width_insert_delete_reaction(
  db_con,
  table_name = "continuity_width"
)
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{table_name}{table name.}
}
\value{
text
}
\description{
Add trigger function to react from continuity_width insert or delete.
}
