% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_elevation_profiles.R
\name{fct_elevation_profiles_insert_delete_reaction}
\alias{fct_elevation_profiles_insert_delete_reaction}
\title{Add trigger function to react from elevation_profiles insert or delete.}
\usage{
fct_elevation_profiles_insert_delete_reaction(
  db_con,
  table_name = "elevation_profiles"
)
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{table_name}{table name.}
}
\value{
text
}
\description{
Add trigger function to react from elevation_profiles insert or delete.
}
