% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_talweg_metrics.R
\name{fct_talweg_metrics_insert_delete_reaction}
\alias{fct_talweg_metrics_insert_delete_reaction}
\title{Add trigger function to react from talweg_metrics insert or delete.}
\usage{
fct_talweg_metrics_insert_delete_reaction(db_con)
}
\arguments{
\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Add trigger function to react from talweg_metrics insert or delete.
}
