% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_triggers.R
\name{fct_update_hydro_swaths_talweg_metrics_id}
\alias{fct_update_hydro_swaths_talweg_metrics_id}
\title{Add trigger function to update talweg_metrics_id for insert in hydro_swaths.}
\usage{
fct_update_hydro_swaths_talweg_metrics_id(db_con)
}
\arguments{
\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Add trigger function to update talweg_metrics_id for insert in hydro_swaths.
}
