% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_triggers.R
\name{fct_update_hydro_swaths_talweg_metrics_id_talweg_metrics}
\alias{fct_update_hydro_swaths_talweg_metrics_id_talweg_metrics}
\title{Add trigger function to update talweg_metrics_id from hydro_swaths for insert or delete in talweg_metrics.}
\usage{
fct_update_hydro_swaths_talweg_metrics_id_talweg_metrics(db_con)
}
\arguments{
\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Add trigger function to update talweg_metrics_id from hydro_swaths for insert or delete in talweg_metrics.
}
