% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_valley_bottom.R
\name{fct_valley_bottom_insert_delete_reaction}
\alias{fct_valley_bottom_insert_delete_reaction}
\title{Add trigger function to react from valley_bottom insert or delete.}
\usage{
fct_valley_bottom_insert_delete_reaction(db_con, table_name = "valley_bottom")
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{table_name}{table name.}
}
\value{
text
}
\description{
Add trigger function to react from valley_bottom insert or delete.
}
