% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_hydro_stations.R
\docType{data}
\name{hydro_stations}
\alias{hydro_stations}
\title{hydro_stations}
\format{
A data frame with 10 rows and 22 variables:
\describe{
\item{ code_station }{  character }
\item{ libelle_station }{  character }
\item{ uri_station }{  character }
\item{ code_departement }{  character }
\item{ libelle_departement }{  character }
\item{ code_commune }{  character }
\item{ libelle_commune }{  character }
\item{ code_region }{  character }
\item{ libelle_region }{  character }
\item{ code_bassin }{  character }
\item{ libelle_bassin }{  character }
\item{ coordonnee_x_station }{  numeric }
\item{ coordonnee_y_station }{  numeric }
\item{ code_projection_station }{  character }
\item{ libelle_projection_station }{  character }
\item{ code_epsg_station }{  character }
\item{ code_cours_eau }{  character }
\item{ libelle_cours_eau }{  character }
\item{ uri_cours_eau }{  character }
\item{ etat_station }{  character }
\item{ date_maj_station }{  character }
\item{ geometry }{  sfc_POINT,sfc }
}
}
\source{
Source
}
\usage{
hydro_stations
}
\description{
Description.
}
\keyword{datasets}
