% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_sites.R
\name{import_hydro_sites}
\alias{import_hydro_sites}
\title{Get hydrologic sites from Hubeau API.}
\usage{
import_hydro_sites()
}
\value{
sf data.frame with hydrometric sites
}
\description{
Get hydrologic sites from Hubeau API.
}
