% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_stations.R
\name{import_hydro_stations}
\alias{import_hydro_stations}
\title{Get hydrologic stations from Hubeau API.}
\usage{
import_hydro_stations(
  url = "https://hubeau.eaufrance.fr/api/v1/ecoulement/stations?format=json"
)
}
\arguments{
\item{url}{text ecoulement stations from hubeau API.}
}
\value{
sf data.frame with hydrologic stations
}
\description{
Get hydrologic stations from Hubeau API.
}
