% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_landcover.R
\docType{data}
\name{landcover}
\alias{landcover}
\title{landcover}
\format{
A data frame with 10890 rows and 9 variables:
\describe{
\item{ swath }{  numeric }
\item{ label }{  character }
\item{ side }{  character }
\item{ area }{  numeric }
\item{ width1 }{  numeric }
\item{ width2 }{  numeric }
\item{ axis }{  numeric }
\item{ measure }{  numeric }
\item{ width_landcover }{  numeric }
}
}
\source{
Source
}
\usage{
landcover
}
\description{
Description.
}
\keyword{datasets}
