% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_landcover_data.R
\docType{data}
\name{landcover_data}
\alias{landcover_data}
\title{landcover_data}
\format{
A data frame with 8118 rows and 6 variables:
\describe{
\item{ X }{  integer }
\item{ measure }{  integer }
\item{ landcover }{  character }
\item{ landcover_area }{  integer }
\item{ landcover_width }{  numeric }
\item{ AXIS }{  integer }
}
}
\source{
Source
}
\usage{
landcover_data
}
\description{
Description.
}
\keyword{datasets}
