% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_view.R
\name{landcover_full_side_view}
\alias{landcover_full_side_view}
\title{Create landuse_area_full_side view}
\usage{
landcover_full_side_view(db_con, view_name = "landcover_area_full_side")
}
\arguments{
\item{db_con}{database connection parameters.}

\item{view_name}{view name.}
}
\value{
text
}
\description{
Create landuse_area_full_side view
}
