% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdotoro-package.R
\docType{package}
\name{mapdotoro-package}
\alias{mapdotoro}
\alias{mapdotoro-package}
\title{mapdotoro: Mapdo Data Processing}
\description{
A package to make the data processing for the mapdoapp shiny application.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LouisManiere/mapdotoro}
  \item Report bugs at \url{https://github.com/LouisManiere/mapdotoro/issues}
}

}
\author{
\strong{Maintainer}: Louis Manière \email{louis.maniere@ens-lyon.fr} (\href{https://orcid.org/0000-0003-0426-2975}{ORCID})

}
\keyword{internal}
