% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_metrics_data.R
\docType{data}
\name{metrics_data}
\alias{metrics_data}
\title{metrics_data}
\format{
A data frame with 906 rows and 17 variables:
\describe{
\item{ X }{  integer }
\item{ measure }{  integer }
\item{ floodplain_slope }{  numeric }
\item{ floodplain_z0 }{  numeric }
\item{ talweg_height_min }{  numeric }
\item{ talweg_height_median }{  numeric }
\item{ talweg_height_is_interpolated }{  integer }
\item{ talweg_length }{  numeric }
\item{ talweg_elevation_min }{  numeric }
\item{ talweg_elevation_median }{  numeric }
\item{ talweg_slope }{  numeric }
\item{ valley_bottom_width }{  numeric }
\item{ active_channel_width }{  numeric }
\item{ valley_bottom_area }{  integer }
\item{ natural_corridor_width }{  numeric }
\item{ connected_corridor_width }{  numeric }
\item{ AXIS }{  integer }
}
}
\source{
Source
}
\usage{
metrics_data
}
\description{
Description.
}
\keyword{datasets}
