% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_network_data.R
\docType{data}
\name{network_data}
\alias{network_data}
\title{network_data}
\format{
A data frame with 843 rows and 10 variables:
\describe{
\item{ num }{  integer }
\item{ ID }{  character }
\item{ AXIS }{  numeric }
\item{ popup }{  character }
\item{ TOPONYME }{  character }
\item{ hasData }{  character }
\item{ GID }{  integer }
\item{ VALUE }{  integer }
\item{ M }{  integer }
\item{ geom }{  sfc_MULTILINESTRING,sfc }
}
}
\source{
Source
}
\usage{
network_data
}
\description{
Description.
}
\keyword{datasets}
