% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_network_landcover_data.R
\docType{data}
\name{network_landcover_data}
\alias{network_landcover_data}
\title{network_landcover_data}
\format{
A data frame with 7587 rows and 14 variables:
\describe{
\item{ num }{  integer }
\item{ ID }{  character }
\item{ AXIS }{  numeric }
\item{ popup }{  character }
\item{ TOPONYME }{  character }
\item{ hasData }{  character }
\item{ GID }{  integer }
\item{ VALUE }{  integer }
\item{ M }{  integer }
\item{ X }{  integer }
\item{ landcover }{  character }
\item{ landcover_area }{  integer }
\item{ landcover_width }{  numeric }
\item{ geom }{  sfc_MULTILINESTRING,sfc }
}
}
\source{
Source
}
\usage{
network_landcover_data
}
\description{
Description.
}
\keyword{datasets}
