% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_view.R
\name{network_metrics_view}
\alias{network_metrics_view}
\title{Create network_metrics view}
\usage{
network_metrics_view(db_con)
}
\arguments{
\item{db_con}{database connection parameters.}
}
\value{
text
}
\description{
Create network_metrics view
}
