% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pg_export_bassin_hydrographique}
\alias{pg_export_bassin_hydrographique}
\title{#' Write t
#'
#' @param table_name
#' @param drop_existing_table
#'
#' @return
#' @export
#'
#' @examples
write_pgtable <- function(table_name,
drop_existing_table){
table_exist <- dbExistsTable(db_con, table_name)
if (table_exist){
if (drop_existing_table){
query <- glue::glue("DROP TABLE {table_name} CASCADE")
dbExecute(db_con, query)
cat(glue::glue("{table_name} has been dropped from the database"), "\n")
st_write(bassin_hydro, db_con, table_name, driver = "PostgreSQL")
}else{
stop("Process stopped because the table exists and drop_existing_table is FALSE.")
}
} else {
st_write(bassin_hydro, db_con, table_name, driver = "PostgreSQL")
}
return(glue::glue("{table_name} has been successfully created"), "\n")
}
Export bassin_hydrographique data to postgreslq table}
\usage{
pg_export_bassin_hydrographique(
  dataset = bassin_hydrographique,
  table_name = "bassin_hydrographique",
  drop_existing_table = FALSE,
  db_con
)
}
\arguments{
\item{dataset}{bassin_hydrographique sf data.frame.}

\item{table_name}{bassin_hydrographique table name.}

\item{drop_existing_table}{if destination table remove with CASCADE.}

\item{db_con}{database connection.}
}
\value{
text
}
\description{
#' Write t
#'
#' @param table_name
#' @param drop_existing_table
#'
#' @return
#' @export
#'
#' @examples
write_pgtable <- function(table_name,
drop_existing_table){
table_exist <- dbExistsTable(db_con, table_name)
if (table_exist){
if (drop_existing_table){
query <- glue::glue("DROP TABLE {table_name} CASCADE")
dbExecute(db_con, query)
cat(glue::glue("{table_name} has been dropped from the database"), "\n")
st_write(bassin_hydro, db_con, table_name, driver = "PostgreSQL")
}else{
stop("Process stopped because the table exists and drop_existing_table is FALSE.")
}
} else {
st_write(bassin_hydro, db_con, table_name, driver = "PostgreSQL")
}
return(glue::glue("{table_name} has been successfully created"), "\n")
}
Export bassin_hydrographique data to postgreslq table
}
