% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pg_export_hubeau}
\alias{pg_export_hubeau}
\title{Export hydrometric stations from hubeau API}
\usage{
pg_export_hubeau(
  url = "https://hubeau.eaufrance.fr/api/v1/ecoulement/stations?format=json",
  table_name = "hydro_stations",
  drop_existing_table = FALSE,
  db_con,
  region_hydrographique_file_path = file.path("data-raw", "raw-datasets",
    "region_hydrographique.gpkg")
)
}
\arguments{
\item{url}{hubeau url for ecoulement stations.}

\item{table_name}{database table name.}

\item{drop_existing_table}{if destination table remove with CASCADE.}

\item{db_con}{DBI connection to database.}

\item{region_hydrographique_file_path}{file path to region hydrographique dataset.}
}
\value{
text
}
\description{
Export hydrometric stations from hubeau API
}
