% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pg_export_hydro_axis}
\alias{pg_export_hydro_axis}
\title{Export hydrologic network axis to database}
\usage{
pg_export_hydro_axis(
  dataset = hydro_axis,
  table_name = "hydro_axis",
  drop_existing_table = FALSE,
  db_con,
  region_hydrographique_file_path = file.path("data-raw", "raw-datasets",
    "region_hydrographique.gpkg")
)
}
\arguments{
\item{dataset}{hydrologic network axis sf data.frame.}

\item{table_name}{database table name.}

\item{drop_existing_table}{if destination table remove with CASCADE.}

\item{db_con}{DBI connection to database.}

\item{region_hydrographique_file_path}{file path to region hydrographique dataset.}
}
\value{
text
}
\description{
Export hydrologic network axis to database
}
