% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pg_export_landcover_area}
\alias{pg_export_landcover_area}
\title{Export landcover area to database}
\usage{
pg_export_landcover_area(
  dataset = landcover_area_prepared,
  table_name = "landcover_area",
  drop_existing_table = FALSE,
  db_con
)
}
\arguments{
\item{dataset}{landcover data.frame.}

\item{table_name}{database table name.}

\item{drop_existing_table}{if destination table remove with CASCADE.}

\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Export landcover area to database
}
