% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pg_export_region_hydrographique}
\alias{pg_export_region_hydrographique}
\title{Export region_hydrographique data to postgreslq table}
\usage{
pg_export_region_hydrographique(
  dataset = region_hydrographique,
  table_name = "region_hydrographique",
  drop_existing_table = FALSE,
  db_con
)
}
\arguments{
\item{dataset}{region_hydrographique sf data.frame.}

\item{table_name}{region_hydrographique table name.}

\item{drop_existing_table}{if destination table remove with CASCADE.}

\item{db_con}{database connection.}
}
\value{
text
}
\description{
Export region_hydrographique data to postgreslq table
}
