% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pg_export_roe}
\alias{pg_export_roe}
\title{Export roe data to postgreslq table}
\usage{
pg_export_roe(
  dataset = roe,
  table_name = "roe",
  drop_existing_table = FALSE,
  db_con,
  region_hydrographique_file_path = file.path("data-raw", "raw-datasets",
    "region_hydrographique.gpkg")
)
}
\arguments{
\item{dataset}{roe sf data.frame.}

\item{table_name}{roe table name.}

\item{drop_existing_table}{if destination table remove with CASCADE.}

\item{db_con}{database connection.}

\item{region_hydrographique_file_path}{file path to region hydrographique dataset.}
}
\value{
text
}
\description{
Export roe data to postgreslq table
}
