% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pg_export_talweg_metrics}
\alias{pg_export_talweg_metrics}
\title{Export talweg metrics to database}
\usage{
pg_export_talweg_metrics(
  dataset = talweg_metrics_prepared,
  table_name = "talweg_metrics",
  drop_existing_table = FALSE,
  db_con
)
}
\arguments{
\item{dataset}{talweg metrics data.frame.}

\item{table_name}{database table name.}

\item{drop_existing_table}{if destination table remove with CASCADE.}

\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Export talweg metrics to database
}
