% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{pivot_continuity_width}
\alias{pivot_continuity_width}
\title{Prepare continuity dataset to have width for each label}
\usage{
pivot_continuity_width(dataset)
}
\arguments{
\item{dataset}{A continuity data.frame.}
}
\value{
data.frame
}
\description{
Each landcover and continuity dataset have several rows for each swath by label and left or right
side. We need to pivot the dataset to have two row for each side, and width labels are summarized in
columns.
}
