% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{pivot_landcover_continuity_area}
\alias{pivot_landcover_continuity_area}
\title{Prepare landcover or continuity dataset to have area for each label.}
\usage{
pivot_landcover_continuity_area(dataset)
}
\arguments{
\item{dataset}{A landuse or continuity data.frame.}
}
\value{
data.frame
}
\description{
Each landcover and continuity dataset have several rows for each swath by label and left or right
side. We need to pivot the dataset to have two row for each side, and area labels are summarized in
columns.
}
