% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bassin_hydrographique.R
\name{prepare_bassin_hydrographique}
\alias{prepare_bassin_hydrographique}
\title{Prepare bassin_hydrographique dataset to database export.}
\usage{
prepare_bassin_hydrographique(dataset = input_bassin_hydrographique)
}
\arguments{
\item{dataset}{sf data.frame bassin_hydrographique dataset.}
}
\value{
sf data.frame
}
\description{
Prepare bassin_hydrographique dataset to database export.
}
