% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_area.R
\name{prepare_continuity_area}
\alias{prepare_continuity_area}
\title{Prepare continuity area to database export.}
\usage{
prepare_continuity_area(dataset = input_continuity)
}
\arguments{
\item{dataset}{data.frame continuity imported.}
}
\value{
data.frame continuity area prepared.
}
\description{
Prepare continuity area to database export.
}
