% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_width.R
\name{prepare_continuity_width}
\alias{prepare_continuity_width}
\title{Prepare continuity width to database export.}
\usage{
prepare_continuity_width(dataset = input_continuity)
}
\arguments{
\item{dataset}{data.frame continuity imported.}
}
\value{
data.frame continuity width prepared.
}
\description{
Prepare continuity width to database export.
}
