% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_elevation_profiles.R
\name{prepare_elevation_profiles}
\alias{prepare_elevation_profiles}
\title{Prepare elevation profiles to database export.}
\usage{
prepare_elevation_profiles(dataset = input_elevation_profiles)
}
\arguments{
\item{dataset}{data.frame elevation profiles imported.}
}
\value{
data.frame elevation profiles prepared.
}
\description{
Prepare elevation profiles to database export.
}
