% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_axis.R
\name{prepare_hydro_axis}
\alias{prepare_hydro_axis}
\title{Prepare hydrologic axis dataset to database export.}
\usage{
prepare_hydro_axis(
  referentiel_hydro_dataset = input_referentiel_hydro,
  hydro_swaths_dataset = hydro_swaths
)
}
\arguments{
\item{referentiel_hydro_dataset}{sf data.frame hydro_axis dataset.}

\item{hydro_swaths_dataset}{sf data.frame hydro_swaths dataset prepared.}
}
\value{
sf data.frame
}
\description{
Prepare hydrologic axis dataset to database export.
}
