% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_stations.R
\name{prepare_hydro_stations}
\alias{prepare_hydro_stations}
\title{Prepare hydrologic stations to database export.}
\usage{
prepare_hydro_stations(
  dataset = input_hydro_stations,
  region_hydro = region_hydrographique
)
}
\arguments{
\item{dataset}{sf data.frame hydrologic stations imported.}

\item{region_hydro}{sf data.frame hydrologic regions to set gid_region.}
}
\value{
sf data.frame hydrologic stations prepared.
}
\description{
Prepare hydrologic stations to database export.
}
