% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_swaths.R
\name{prepare_hydro_swaths_and_axis}
\alias{prepare_hydro_swaths_and_axis}
\title{Prepare hydro_swaths dataset to database export.}
\usage{
prepare_hydro_swaths_and_axis(
  swaths_dataset = input_swaths,
  referentiel_hydro_dataset = input_referentiel_hydro,
  region_hydro = region_hydrographique
)
}
\arguments{
\item{swaths_dataset}{sf data.frame swaths dataset.}

\item{referentiel_hydro_dataset}{sf data.frame hydrographic network dataset.}

\item{region_hydro}{sf data.frame region_hydrographique dataset prepared.}
}
\value{
list with two sf data.frame
}
\description{
Prepare hydro_swaths dataset to database export.
}
