% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_landcover.R
\name{prepare_landcover_area}
\alias{prepare_landcover_area}
\title{Prepare landcover area to database export.}
\usage{
prepare_landcover_area(dataset = input_landcover)
}
\arguments{
\item{dataset}{data.frame landcover imported.}
}
\value{
data.frame landcover area prepared.
}
\description{
Prepare landcover area to database export.
}
