% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{prepare_landcover_continuity_area}
\alias{prepare_landcover_continuity_area}
\title{Prepare landcover or continuity dataset to have area for each label}
\usage{
prepare_landcover_continuity_area(dataset)
}
\arguments{
\item{dataset}{A landuse or continuity data.frame.}
}
\value{
data.frame
}
\description{
Prepare landcover or continuity dataset to have area for each label
}
\examples{
landcover_prepared <- prepare_landcover_continuity_area(landcover)
}
