% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_dataset.R
\name{prepare_region_hydrographique}
\alias{prepare_region_hydrographique}
\title{Prepare region_hydrographique dataset to database export.}
\usage{
prepare_region_hydrographique(dataset = input_region_hydrographique)
}
\arguments{
\item{dataset}{region_hydrographique sf data.frame}
}
\value{
sf data.frame
}
\description{
Prepare region_hydrographique dataset to database export.
}
