% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_dataset.R
\name{prepare_roe}
\alias{prepare_roe}
\title{Prepare roe dataset to database export.}
\usage{
prepare_roe(dataset = input_roe)
}
\arguments{
\item{dataset}{roe sf data.frame}
}
\value{
sf data.frame
}
\description{
Prepare roe dataset to database export.
}
