% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_talweg_metrics.R
\name{prepare_talweg_metrics}
\alias{prepare_talweg_metrics}
\title{Prepare talweg metrics to database export.}
\usage{
prepare_talweg_metrics(dataset = input_talweg_metrics)
}
\arguments{
\item{dataset}{data.frame talweg metrics imported.}
}
\value{
data.frame talweg metrics prepared.
}
\description{
Prepare talweg metrics to database export.
}
