% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_valley_bottom.R
\name{prepare_valley_bottom}
\alias{prepare_valley_bottom}
\title{Prepare valley_bottom to database export.}
\usage{
prepare_valley_bottom(dataset = input_valley_bottom)
}
\arguments{
\item{dataset}{data.frame valley_bottom imported.}
}
\value{
data.frame valley bottom prepared.
}
\description{
Prepare valley_bottom to database export.
}
