% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove rows in database table based on field identifier.}
\usage{
remove_rows(dataset, field_identifier, table_name)
}
\arguments{
\item{dataset}{sf data.frame dataset.}

\item{field_identifier}{text field identifier name to identified rows to remove.}

\item{table_name}{text database table name.}
}
\value{
text number of row deleted.
}
\description{
Remove rows in database table based on field identifier.
}
