% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R, R/fct_upsert_db.R
\name{set_displayed_bassin_region}
\alias{set_displayed_bassin_region}
\title{Set display column value for bassin or region table.}
\usage{
set_displayed_bassin_region(table_name, code_with_data)

set_displayed_bassin_region(table_name, code_with_data)
}
\arguments{
\item{table_name}{bassin or region table name.}

\item{code_with_data}{A vector with the list of cdbh for bassin, cdregionhy for region value to set the displayed polygons.}

\item{code_bassin_with_data}{A vector with the list of cdbh value to set the displayed bassin.}
}
\value{
text

text
}
\description{
Set display column value for bassin or region table.

Set display column value for bassin or region table.
}
