% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_swaths.R
\docType{data}
\name{swaths}
\alias{swaths}
\title{swaths}
\format{
A data frame with 779 rows and 6 variables:
\describe{
\item{ GID }{  numeric }
\item{ AXIS }{  numeric }
\item{ VALUE }{  integer }
\item{ M }{  numeric }
\item{ DRAINAGE }{  numeric }
\item{ geometry }{  sfc_MULTIPOLYGON,sfc }
}
}
\source{
Source
}
\usage{
swaths
}
\description{
Description.
}
\keyword{datasets}
