% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_talweg_metrics.R
\docType{data}
\name{talweg_metrics}
\alias{talweg_metrics}
\title{talweg_metrics}
\format{
A data frame with 605 rows and 9 variables:
\describe{
\item{ swath }{  numeric }
\item{ elevation_talweg }{  numeric }
\item{ elevation_talweg_med }{  numeric }
\item{ height_valley_bottom }{  numeric }
\item{ slope_talweg }{  numeric }
\item{ slope_valley_bottom }{  numeric }
\item{ axis }{  numeric }
\item{ measure }{  numeric }
\item{ sinuosity }{  numeric }
}
}
\source{
Source
}
\usage{
talweg_metrics
}
\description{
Description.
}
\keyword{datasets}
