% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_area.R
\name{trig_continuity_area}
\alias{trig_continuity_area}
\title{Create trigger to update tables from continuity_area modifications.}
\usage{
trig_continuity_area(db_con, table_name = "continuity_area")
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{table_name}{table name.}
}
\value{
text
}
\description{
Create trigger to update tables from continuity_area modifications.
}
