% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_width.R
\name{trig_continuity_width}
\alias{trig_continuity_width}
\title{Create trigger to update tables from continuity_width modifications.}
\usage{
trig_continuity_width(db_con, table_name = "continuity_width")
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{table_name}{table name.}
}
\value{
text
}
\description{
Create trigger to update tables from continuity_width modifications.
}
