% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_elevation_profiles.R
\name{trig_elevation_profiles}
\alias{trig_elevation_profiles}
\title{Create trigger to update tables from elevation_profiles modifications.}
\usage{
trig_elevation_profiles(db_con, table_name = "elevation_profiles")
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{table_name}{table name.}
}
\value{
text
}
\description{
Create trigger to update tables from elevation_profiles modifications.
}
