% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_swaths.R
\name{trig_hydro_swaths}
\alias{trig_hydro_swaths}
\title{Create trigger to update tables from hydro_swaths modifications.}
\usage{
trig_hydro_swaths(db_con)
}
\arguments{
\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Create trigger to update tables from hydro_swaths modifications.
}
