% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_landcover_area.R
\name{trig_landcover_area}
\alias{trig_landcover_area}
\title{Create trigger to update tables from landcover_area modifications.}
\usage{
trig_landcover_area(db_con)
}
\arguments{
\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Create trigger to update tables from landcover_area modifications.
}
