% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_triggers.R
\name{trig_update_talweg_metrics_id}
\alias{trig_update_talweg_metrics_id}
\title{Create trigger to update talweg_metrics_id in hydro_swaths table.}
\usage{
trig_update_talweg_metrics_id(db_con)
}
\arguments{
\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Create trigger to update talweg_metrics_id in hydro_swaths table.
}
