% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_valley_bottom.R
\name{trig_valley_bottom}
\alias{trig_valley_bottom}
\title{Create trigger to update tables from valley_bottom modifications.}
\usage{
trig_valley_bottom(db_con, table_name = "valley_bottom")
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{table_name}{table name.}
}
\value{
text
}
\description{
Create trigger to update tables from valley_bottom modifications.
}
