% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_troncon_bdtopo_id.R
\docType{data}
\name{troncon_bdtopo_id}
\alias{troncon_bdtopo_id}
\title{troncon_bdtopo_id}
\format{
A data frame with 536 rows and 2 variables:
\describe{
\item{ id_troncon }{  character }
\item{ axis }{  character }
}
}
\source{
Source
}
\usage{
troncon_bdtopo_id
}
\description{
Description.
}
\keyword{datasets}
