% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bassin_hydrographique.R
\name{upsert_bassin_hydrographique}
\alias{upsert_bassin_hydrographique}
\title{Delete existing rows and insert hydrologic bassin to database}
\usage{
upsert_bassin_hydrographique(
  dataset = bassin_hydrographique,
  table_name = "bassin_hydrographique",
  db_con,
  field_identifier = "cdbh"
)
}
\arguments{
\item{dataset}{sf data.frame hydrologic bassin.}

\item{table_name}{text database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrologic bassin to database
}
