% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_continuity_width.R
\name{upsert_continuity_width}
\alias{upsert_continuity_width}
\title{Delete existing rows and insert continuity width to database.}
\usage{
upsert_continuity_width(
  dataset = continuity_width,
  table_name = "continuity_width",
  db_con,
  field_identifier = "axis"
)
}
\arguments{
\item{dataset}{sf data.frame continuity width.}

\item{table_name}{text database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert continuity width to database.
}
