% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_axis.R
\name{upsert_hydro_axis}
\alias{upsert_hydro_axis}
\title{Delete existing rows and insert hydrologic axis to database.}
\usage{
upsert_hydro_axis(
  dataset = hydro_axis,
  table_name = "hydro_axis",
  db_con,
  field_identifier = "axis"
)
}
\arguments{
\item{dataset}{sf data.frame hydrologic axis.}

\item{table_name}{database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrologic axis to database.
}
