% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_sites.R
\name{upsert_hydro_sites}
\alias{upsert_hydro_sites}
\title{Delete existing rows and insert hydrometric sites to database.}
\usage{
upsert_hydro_sites(
  dataset = hydro_sites,
  table_name = "hydro_sites",
  db_con,
  field_identifier = "code_site"
)
}
\arguments{
\item{dataset}{sf data.frame hydrometric sites}

\item{table_name}{text database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrometric sites to database.
}
