% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_stations.R
\name{upsert_hydro_stations}
\alias{upsert_hydro_stations}
\title{Delete existing rows and insert hydrologic stations to database.}
\usage{
upsert_hydro_stations(
  dataset = hydro_stations,
  table_name = "hydro_stations",
  db_con,
  field_identifier = "code_station"
)
}
\arguments{
\item{dataset}{sf data.frame hydrologic stations.}

\item{table_name}{text database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrologic stations to database.
}
