% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_swaths.R
\name{upsert_hydro_swaths}
\alias{upsert_hydro_swaths}
\title{Delete existing rows and insert hydrologic network splited by swaths to database.}
\usage{
upsert_hydro_swaths(
  dataset = hydro_swaths,
  table_name = "hydro_swaths",
  db_con,
  field_identifier = "axis"
)
}
\arguments{
\item{dataset}{sf data.frame hydrologic network splited.}

\item{table_name}{database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrologic network splited by swaths to database.
}
