% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_swaths.R
\name{upsert_hydro_swaths_and_axis}
\alias{upsert_hydro_swaths_and_axis}
\title{Delete existing rows and insert hydrologic network splited by swaths to database.}
\usage{
upsert_hydro_swaths_and_axis(
  hydro_swaths_dataset = hydro_swaths,
  hydro_swaths_table_name = "hydro_swaths",
  hydro_axis_dataset = hydro_swaths,
  hydro_axis_table_name = "hydro_axis",
  db_con,
  field_identifier = "axis"
)
}
\arguments{
\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}

\item{dataset}{sf data.frame hydrologic network splited.}

\item{table_name}{database table name.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrologic network splited by swaths to database.
}
