% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_landcover_area.R
\name{upsert_landcover_area}
\alias{upsert_landcover_area}
\title{Delete existing rows and insert landcover area to database.}
\usage{
upsert_landcover_area(
  dataset = landcover_area,
  table_name = "landcover_area",
  db_con,
  field_identifier = "axis"
)
}
\arguments{
\item{dataset}{sf data.frame landcover area.}

\item{table_name}{text database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert landcover area to database.
}
