% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upsert_db.R
\name{upsert_region_hydrographique}
\alias{upsert_region_hydrographique}
\title{Delete existing rows and insert hydrologic region to database}
\usage{
upsert_region_hydrographique(
  dataset = region_hydrographique,
  table_name = "region_hydrographique",
  db_con,
  field_identifier = "cdregionhy"
)
}
\arguments{
\item{dataset}{sf data.frame hydrologic region.}

\item{table_name}{database table name.}

\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrologic region to database
}
