% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_roe.R
\name{upsert_roe}
\alias{upsert_roe}
\title{Delete existing rows and insert roe to database}
\usage{
upsert_roe(
  dataset = roe,
  table_name = "roe",
  db_con,
  field_identifier = "cdobstecou"
)
}
\arguments{
\item{dataset}{sf data.frame roe.}

\item{table_name}{text database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert roe to database
}
