% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_talweg_metrics.R
\name{upsert_talweg_metrics}
\alias{upsert_talweg_metrics}
\title{Delete existing rows and insert talweg metrics to database.}
\usage{
upsert_talweg_metrics(
  dataset = talweg_metrics,
  table_name = "talweg_metrics",
  db_con,
  field_identifier = "axis"
)
}
\arguments{
\item{dataset}{sf data.frame talweg metrics.}

\item{table_name}{text database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert talweg metrics to database.
}
