% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_valley_bottom.R
\name{upsert_valley_bottom}
\alias{upsert_valley_bottom}
\title{Delete existing rows and insert valley botom to database.}
\usage{
upsert_valley_bottom(
  dataset = valley_bottom,
  table_name = "valley_bottom",
  db_con,
  field_identifier = "axis"
)
}
\arguments{
\item{dataset}{sf data.frame valley bottom.}

\item{table_name}{text database table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert valley botom to database.
}
