% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_valley_bottom.R
\docType{data}
\name{valley_bottom}
\alias{valley_bottom}
\title{valley_bottom}
\format{
A data frame with 1210 rows and 7 variables:
\describe{
\item{ swath }{  numeric }
\item{ side }{  character }
\item{ area }{  numeric }
\item{ width1 }{  numeric }
\item{ width2 }{  numeric }
\item{ axis }{  numeric }
\item{ measure }{  numeric }
}
}
\source{
Source
}
\usage{
valley_bottom
}
\description{
Description.
}
\keyword{datasets}
