% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_datasets.R
\name{check_duplicate}
\alias{check_duplicate}
\title{Check for duplicate measure in streams axis.}
\usage{
check_duplicate(dataset, axis_field = "AXIS", measure_field = "M")
}
\arguments{
\item{dataset}{sf data.frame.}

\item{axis_field}{axis stream id.}

\item{measure_field}{axis stream measure from exutoire.}
}
\value{
sf data.frame with all duplicated rows.
}
\description{
Check for duplicate measure in streams axis.
}
\examples{
check_duplicate(swaths)
}
