% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_datasets.R
\name{clean_duplicated}
\alias{clean_duplicated}
\title{Drop all the duplicated rows from a duplicated sf data.frame.}
\usage{
clean_duplicated(
  dataset,
  duplicated_dataset,
  axis_field = "AXIS",
  measure_field = "M"
)
}
\arguments{
\item{dataset}{sf data.frame.}

\item{duplicated_dataset}{sf data.frame.}

\item{axis_field}{axis stream id.}

\item{measure_field}{axis stream measure from exutoire.}
}
\value{
sf data.frame.
}
\description{
Drop all the duplicated rows from a duplicated sf data.frame.
}
\examples{
clean_duplicated(swaths,
                 check_duplicate(swaths))
}
