% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_landcover_area.R
\name{fct_landcover_area_insert_delete_reaction}
\alias{fct_landcover_area_insert_delete_reaction}
\title{Add trigger function to react from landcover_area insert or delete.}
\usage{
fct_landcover_area_insert_delete_reaction(db_con)
}
\arguments{
\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Add trigger function to react from landcover_area insert or delete.
}
