% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pg_export_bassin_hydrographique}
\alias{pg_export_bassin_hydrographique}
\title{Export bassin_hydrographique data to postgreslq table}
\usage{
pg_export_bassin_hydrographique(
  dataset = bassin_hydrographique,
  table_name = "bassin_hydrographique",
  drop_existing_table = FALSE,
  db_con
)
}
\arguments{
\item{dataset}{bassin_hydrographique sf data.frame.}

\item{table_name}{bassin_hydrographique table name.}

\item{drop_existing_table}{if destination table remove with CASCADE.}

\item{db_con}{database connection.}
}
\value{
text
}
\description{
Export bassin_hydrographique data to postgreslq table
}
