% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{pivot_continuity_width}
\alias{pivot_continuity_width}
\title{Prepare continuity dataset to have width for each label}
\usage{
pivot_continuity_width(dataset)
}
\arguments{
\item{dataset}{A continuity data.frame.}
}
\value{
data.frame
}
\description{
Prepare continuity dataset to have width for each label
}
\examples{
landcover_prepared <- pivot_continuity_width(input_continuity)
}
