% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_dataset.R
\name{prepare_roe}
\alias{prepare_roe}
\title{Prepare roe dataset to database export.}
\usage{
prepare_roe(dataset = input_roe, region_hydro = region_hydrographique)
}
\arguments{
\item{dataset}{sf data.frame roe.}

\item{region_hydro}{sf data.frame hydrographic regions to set spatial join on gid_region.}
}
\value{
sf data.frame
}
\description{
Prepare roe dataset to database export.
}
