% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{set_displayed_bassin_region}
\alias{set_displayed_bassin_region}
\title{Set display column value for bassin or region table.}
\usage{
set_displayed_bassin_region(
  table_name,
  display_codes_bassin_or_region,
  field_identifier,
  db_con
)
}
\arguments{
\item{table_name}{bassin or region table name.}

\item{display_codes_bassin_or_region}{A vector with the list of cdbh for bassin, cdregionhy for region value to set the displayed polygons.}

\item{field_identifier}{text field identifier name to identified rows to remove.}

\item{db_con}{DBI connection to database.}
}
\value{
text
}
\description{
Set display column value for bassin or region table.
}
