% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{set_displayed_bassin_region}
\alias{set_displayed_bassin_region}
\title{Set display column value for bassin or region table.}
\usage{
set_displayed_bassin_region(table_name, displayed_gid)
}
\arguments{
\item{table_name}{bassin or region table name.}

\item{display_gid_bassin}{A numeric vector with gid bassin value.}
}
\value{
text
}
\description{
Set display column value for bassin or region table.
}
