% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_swaths.R
\name{upsert_hydro_swaths}
\alias{upsert_hydro_swaths}
\title{Delete existing rows and insert hydrologic network splited by swaths to database}
\usage{
upsert_hydro_swaths(
  dataset = hydro_swaths_prepared,
  table_name = "hydro_swaths",
  db_con
)
}
\arguments{
\item{dataset}{hydrologic network splited sf data.frame.}

\item{table_name}{database table name.}

\item{db_con}{DBI connection to database.}

\item{region_hydrographique_file_path}{file path to region hydrographique dataset.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrologic network splited by swaths to database
}
