% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydro_swaths.R
\name{upsert_hydro_swaths_and_axis}
\alias{upsert_hydro_swaths_and_axis}
\title{Delete existing rows and insert hydrologic network splited by swaths to database.}
\usage{
upsert_hydro_swaths_and_axis(
  hydro_swaths_dataset = hydro_swaths,
  hydro_swaths_table_name = "hydro_swaths",
  hydro_axis_dataset = hydro_axis,
  hydro_axis_table_name = "hydro_axis",
  db_con,
  field_identifier = "axis"
)
}
\arguments{
\item{hydro_swaths_dataset}{sf data.frame hydro swaths prepared.}

\item{hydro_swaths_table_name}{text hydro_swaths table name.}

\item{hydro_axis_dataset}{sf data.frame hydro axis prepared.}

\item{hydro_axis_table_name}{text hydro_axis table name.}

\item{db_con}{DBI connection to database.}

\item{field_identifier}{text field identifier name to identified rows to remove.}
}
\value{
text
}
\description{
Delete existing rows and insert hydrologic network splited by swaths to database.
}
